Cosmetic rename textswapper.
